requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
openai>=0.27.0
sentence-transformers>=2.2.2
numpy>=1.23.5
pandas>=1.5.3
faiss-cpu>=1.7.4  # or faiss-gpu if you have GPU support
pinecone # For Pinecone vector database
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
openai>=0.27.0
sentence-transformers>=2.2.2
numpy>=1.23.5
pandas>=1.5.3
faiss-cpu>=1.7.4  # or faiss-gpu if you have GPU support
pinecone # For Pinecone vector database
streamlit>=1.29.0
requests>=2.28.1